App.py restructuring
Offloads some of the initializers on the app.py file here
Leave app.py purely for the routes
Reduces the size of the app.py file
Allows for easy skimming of app.py
